(Finally) implemented a depth-first draft
There's a lot of strange behaviour, but the algorithm mostly works remarkably well. Getting it working was sort of like pulling teeth: had to store the gold list internally with link because the normal method is bugged; realised after a lot of brain-wracking I need to make a new path AFTER moving onto gold or I mess up the pathfinding; need to do a lot of .x and .y because for some reason Pose() HATES comparisons of any kind. Still, Link loves to throw himself into pits and wumpuses at seemingly random on a bad test. The challenge now is working out why that is. Still, a very satisfying start.