Attempted to fix the draft's bugs - mixed results
Worst consistent issues are that the pathfinding, despite the fact it SHOULD NOT be doing this, continues to rarely run into pits, and more frequently run into wumpuses. I'm baffled by this, and wondering if the root cause is that the built-in methods for detecting both are bugged. I'm not moving on until I fix it (unfortunately) so I'm gonna just have to do rigorous bug testing.
In other notes, I realised that the way I was handling gold updates was bad, and added a new variable to tell when gold has gone down.