Added algorithm selector & fixed weird bug
For whatever reason adding the algorithm selector totally broke the way Pose generation was handled. Instead of correctly creating a new Pose it decided it was going to have all the Poses share the same memory address which totally broke the code. No clue why this happened - I assume it's some weird quirk of Python. Made it explicitly attach a new Pose address each loop, which should prevent the issue coming up again in the future