Implemented breadth-first search
Wow that was way faster. I had a theory that hashing out most of the teething problems implementing depth-first meant subsequent implementations would be a lot smoother. Did not expect it to this extent! That was, what, an hour?
Pulled the validMoves loop out of depth-first and made it it's own method that both depth-first and breadth-first can call. I imagine other algorithms will use it for that too