Added better dynamic wumpus detection
Previously it would check adjacent to Link to see if a Wumpus was around it and re-run the algorithm if so, but I realised this doesn't cover cases where the Wumpus will move onto the next position Link moves to. Wrote new code to instead check the next position. This VASTLY improves consistency!